Extract into recommendation service
